Refactor: move timestamp functions into their own file
