+++ sno/commit.py
@@ -20,6 +20,7 @@ from .status import (
     get_diff_status_json,
     diff_status_to_text,
 )
+from .timestamps import to_iso8601_utc, to_iso8601_tz, commit_time_to_text
 from .working_copy import WorkingCopy
 from .structure import RepositoryStructure
 from .cli_util import MutexOption, do_json_option
@@ -200,25 +201,3 @@ def commit_json_to_text(jdict):
     diff = diff_status_to_text(jdict["changes"])
     datetime = commit_time_to_text(jdict["commitTime"], jdict["commitTimeOffset"])
     return f"[{branch} {commit}] {message}\n{diff}\n  Date: {datetime}"
-
-
-def to_iso8601_utc(datetime):
-    """Returns a string like: 2020-03-26T09:10:11Z"""
-    isoformat = datetime.astimezone(timezone.utc).replace(tzinfo=None).isoformat()
-    return f"{isoformat}Z"
-
-
-def to_iso8601_tz(timedelta):
-    """Returns a string like "+05:00" or "-05:00" (ie five hours ahead or behind)."""
-    abs_delta = datetime.utcfromtimestamp(abs(timedelta).seconds).strftime('%H:%M')
-    return f"+{abs_delta}" if abs(timedelta) == timedelta else f"-{abs_delta}"
-
-
-def commit_time_to_text(iso8601z, iso_offset):
-    """
-    Given an isoformat time in UTC, and a isoformat timezone offset,
-    returns the time in a human readable format, for that timezone.
-    """
-    right_time = datetime.fromisoformat(iso8601z.replace("Z", "+00:00"))
-    right_tzinfo = datetime.fromisoformat(iso8601z.replace("Z", iso_offset))
-    return right_time.astimezone(right_tzinfo.tzinfo).strftime("%c %z")
